refactor(Referee): 메소드 분리 및 패키지 이동
before:
java/racinggame

after:
java/domain
